[ALL] remove warnning

+++ dtaudio/dtaudio.c
@@ -129,7 +129,7 @@ void audio_update_pts(void *priv)
     if (delay_pts == -1) {
         delay_pts = 0;
     }
-    dt_debug(TAG, "[%s:%d] pts:%lld  delay:%lld  actual:%lld time:%d cur_time:%lld\n", __FUNCTION__, __LINE__, pts, delay_pts, pts - delay_pts, (pts - delay_pts) / 90000, dt_gettime());
+    dt_debug(TAG, "[%s:%d] pts:%lld  delay:%lld  actual:%lld time:%d cur_time:%lld\n", __FUNCTION__, __LINE__, pts, delay_pts, pts - delay_pts, (int)((pts - delay_pts) / 90000), dt_gettime());
     if (delay_pts < pts) {
         pts -= delay_pts;
     } else {
